Add Titanfall 2 (draft)
This config is able to launch and load the Northstar launcher, but fails on a D3D issue. I have yet to determine the cause. I'm hoping future versions of the launcher fix the issue as they have a focus on improving the D3D calls in their GitHub repo.