Added fix for connection timeout error
Set the ConfigFormat to use quotes around all settings. I believe this was causing games with spaces in the name to be non-joinable. The strange part is you could set it back to having a space and it would work going forward.

Also set the ExitMethod to String and quit. The user must type yes following since there is no option to do two input commands to quit a server. CtrlC doesn't work in AMP on Windows for this game. And CtrlC doesn't work for this game on the Linux Terminal. OS_CLOSE ended in a kill.

Removed the config file and used the start/wait/stop update stages instead.