ConfigFormat changes not needed
I believe that this game's dedicated server is the cause of the timeout issue. It is still in an early beta stage at this point. I have tried to recreate the problem and no matter what, I cannot now. After the first initial connection to any server it works from then on. The ini file is intended to not have quotes in this case given the examples at the top. There are other settings I've put in the GUI but do not seem to function in the server yet, things like "No Food", "No Death", etc. have no impact. You will die and you will be angry. Please contact the develop to have them fix their game.