Fix Risk of Rain 2
The devs broke the log. It no longer shows player join/leave properly. It just has ??? for most values. This can't be fixed, but the join regex can be. They added a space before the line previously, now they added one after the line. I'm adding buffers to each line as a failsafe.