Add 7 Days to Die Generic (#112)
* Add 7 Days to Die Generic

This is the draft of the 7 Days to Die generic config. This is not thoroughly tested yet, and not tried on Linux at all. I believe most everything works as expected including UMOD in Windows though.

Known Issues:
1. No Telnet connection due to related feature request: https://github.com/CubeCoders/AMP/issues/623
2. No FPS Status - This is not currently possible with Generic

* Multiple changes

-Fix Telnet
-Move Update Stages to separate file
-Setup for pre-20.4 config file option

* Fix my dummy mistakes

It's better now. It's much better now.

* Update Join Regex

This accounts for the old or new version's regex so it should work on any version.

* Move Update Settings

Moved the Release Stream and UMOD to the SteamCMD and Updates tab.