Stationeers: convert to legacy server download
The new server build does not work with the current AMP template. Converting to the `previous` branch so that the legacy server build is downloaded, pending an updated AMP template being committed to cater for the new server build (which is pending the Stationeers devs fixing their server console).

App ready regex also fixed.