Updated Download and Copy settings.
Changed the UpdateSourceData to a more reliable URL to accommodate for the authors GitLab main/master branch switching.

Added UpdateSourceArgs to rename the download file to .zip as it was coming down as dl?v=exp or dl?=stable

Adjusted the copy commands to accommodate for the extracted folder name changes between -main and -master.