Update GatekeeperV2
This removes the requirement to use a special Docker image since the python3 tagged ampbase image includes pip now. This also adds the additional stuff into a script that runs with the update stages instead of being handled by the before mentioned image. These are very quick installs that take only a couple of seconds.