Add Sons of the Forest
This adds Sons of the Forest. There are no known limitations, but one bug with the game server that if the ports aren't declared as command line args, it fails to start since the ports are not able to be read from the config file. This template caters to both so it should not be an issue.