Add Sons of the Forest (#406)
* Add Sons of the Forest

This adds Sons of the Forest. There are no known limitations, but one bug with the game server that if the ports aren't declared as command line args, it fails to start since the ports are not able to be read from the config file. This template caters to both so it should not be an issue.

* Update sons-of-the-forestupdates.json

* Update sons-of-the-forest.kvp

* Update sons-of-the-forestconfig.json

* Update sons-of-the-forest.kvp

* Update sons-of-the-forestconfig.json

* Update sons-of-the-forestconfig.json

* Update sons-of-the-forestconfig.json

* Update sons-of-the-forestconfig.json

* Update sons-of-the-forestconfig.json

* Update sons-of-the-forestconfig.json

* Update sons-of-the-forestconfig.json

---------

Co-authored-by: Greelan <53196309+Greelan@users.noreply.github.com>