Adds exclusivejoin and NoTransferFromFiltering
This Adds NoTransferFromFiltering = Prevents ARK Data usage between single player and servers who do not have a cluster ID. Even with a cluster id set, it is suggested to keep this option to completely disable code path allowing it.

And Exclusivejoin = Activate a whitelist only mode on the server, allowing players join if added to the allow lists. This option uses PlayersExclusiveJoinList.txt and PlayersJoinNoCheckList.txt files. The files are located in /ShooterGame/Binaries/<PLATFORM>/ where <PLATFORM> is either Linux or Win64 depending on host operating system. PlayersExclusiveJoinList.txt is used straightforward, allowing players only whose IDs are present in the file to connect in the server.