Major Rewrite
Now updates do not require downloading the whole repository to be downloaded. Some commands have been shortened for readability and and tested on both Windows and Linux to work. It was not a completely clean install of Windows nor Linux so someone else may need to test the module to see if any problems may come up.