ARK:SA Unified Save Data (usestore)
This adds support for usestore as well as converttostore and BackupTransferPlayerDatas which is beneficial for clustering and swapping between servers

Per Patch Notes:
Additional Server related params/config options for using a “usestore” save system which combines player profiles, tribe data, and the world save into one single file. Note: This is not new, but not previously documented to our knowledge and what we use on our Official Network.
-converttostore (imports the current player and tribe data into the .ark save file)
Run with -converttostore -usestore and then the next time you relaunch, you should not need -converttostore as all the data will be integrated.
Important to note: .arkprofile & .arktribe files are not automatically deleted after conversion so should be deleted manually if you want to save storage space.
-usestore (utilizes a different saving method that stores player and tribe data inside of save files)
Important to note: when rolling back save data simply do as normal and all the arkprofile and arktribe data will be in sync with the backup save already.